[ML] Fix date picker for the Log pattern analysis (#143684)
* url state provider

* update url state provider for explain_log_rate_spikes_app_state.tsx

(cherry picked from commit f61dec4c3d0ce3ca3669c1bcb40adced259b93a7)
