+++ x-pack/plugins/aiops/public/components/explain_log_rate_spikes/explain_log_rate_spikes_app_state.tsx
@@ -5,11 +5,7 @@
  * 2.0.
  */
 
-import React, { FC, useCallback } from 'react';
-import { parse, stringify } from 'query-string';
-import { isEqual } from 'lodash';
-import { encode } from 'rison-node';
-import { useHistory, useLocation } from 'react-router-dom';
+import React, { FC } from 'react';
 
 import { EuiCallOut } from '@elastic/eui';
 
@@ -24,15 +20,7 @@ import {
   SearchQueryLanguage,
   SavedSearchSavedObject,
 } from '../../application/utils/search_utils';
-import {
-  Accessor,
-  Dictionary,
-  parseUrlState,
-  Provider as UrlStateContextProvider,
-  isRisonSerializationRequired,
-  getNestedProperty,
-  SetUrlState,
-} from '../../hooks/use_url_state';
+import { UrlStateProvider } from '../../hooks/use_url_state';
 import type { AiopsAppDependencies } from '../../hooks/use_aiops_app_context';
 import { AiopsAppContext } from '../../hooks/use_aiops_app_context';
 
@@ -77,71 +65,6 @@ export const ExplainLogRateSpikesAppState: FC<ExplainLogRateSpikesAppStateProps>
   savedSearch,
   appDependencies,
 }) => {
-  const history = useHistory();
-  const { search: urlSearchString } = useLocation();
-
-  const setUrlState: SetUrlState = useCallback(
-    (
-      accessor: Accessor,
-      attribute: string | Dictionary<any>,
-      value?: any,
-      replaceState?: boolean
-    ) => {
-      const prevSearchString = urlSearchString;
-      const urlState = parseUrlState(prevSearchString);
-      const parsedQueryString = parse(prevSearchString, { sort: false });
-
-      if (!Object.prototype.hasOwnProperty.call(urlState, accessor)) {
-        urlState[accessor] = {};
-      }
-
-      if (typeof attribute === 'string') {
-        if (isEqual(getNestedProperty(urlState, `${accessor}.${attribute}`), value)) {
-          return prevSearchString;
-        }
-
-        urlState[accessor][attribute] = value;
-      } else {
-        const attributes = attribute;
-        Object.keys(attributes).forEach((a) => {
-          urlState[accessor][a] = attributes[a];
-        });
-      }
-
-      try {
-        const oldLocationSearchString = stringify(parsedQueryString, {
-          sort: false,
-          encode: false,
-        });
-
-        Object.keys(urlState).forEach((a) => {
-          if (isRisonSerializationRequired(a)) {
-            parsedQueryString[a] = encode(urlState[a]);
-          } else {
-            parsedQueryString[a] = urlState[a];
-          }
-        });
-        const newLocationSearchString = stringify(parsedQueryString, {
-          sort: false,
-          encode: false,
-        });
-
-        if (oldLocationSearchString !== newLocationSearchString) {
-          const newSearchString = stringify(parsedQueryString, { sort: false });
-          if (replaceState) {
-            history.replace({ search: newSearchString });
-          } else {
-            history.push({ search: newSearchString });
-          }
-        }
-      } catch (error) {
-        // eslint-disable-next-line no-console
-        console.error('Could not save url state', error);
-      }
-    },
-    [history, urlSearchString]
-  );
-
   if (!dataView) return null;
 
   if (!dataView.isTimeBased()) {
@@ -165,11 +88,11 @@ export const ExplainLogRateSpikesAppState: FC<ExplainLogRateSpikesAppStateProps>
 
   return (
     <AiopsAppContext.Provider value={appDependencies}>
-      <UrlStateContextProvider value={{ searchString: urlSearchString, setUrlState }}>
+      <UrlStateProvider>
         <SpikeAnalysisTableRowStateProvider>
           <ExplainLogRateSpikesPage dataView={dataView} savedSearch={savedSearch} />
         </SpikeAnalysisTableRowStateProvider>
-      </UrlStateContextProvider>
+      </UrlStateProvider>
     </AiopsAppContext.Provider>
   );
 };
