+++ x-pack/plugins/aiops/public/hooks/use_url_state.tsx
@@ -5,9 +5,12 @@
  * 2.0.
  */
 
-import { parse } from 'query-string';
+import React, { FC } from 'react';
+import { parse, stringify } from 'query-string';
 import { createContext, useCallback, useContext, useMemo } from 'react';
-import { decode } from 'rison-node';
+import { decode, encode } from 'rison-node';
+import { useHistory, useLocation } from 'react-router-dom';
+import { isEqual } from 'lodash';
 
 export interface Dictionary<TValue> {
   [id: string]: TValue;
@@ -86,6 +89,79 @@ export const aiopsUrlStateStore = createContext<UrlState>({
 });
 
 export const { Provider } = aiopsUrlStateStore;
+
+export const UrlStateProvider: FC = ({ children }) => {
+  const { Provider: StateProvider } = aiopsUrlStateStore;
+
+  const history = useHistory();
+  const { search: urlSearchString } = useLocation();
+
+  const setUrlState: SetUrlState = useCallback(
+    (
+      accessor: Accessor,
+      attribute: string | Dictionary<any>,
+      value?: any,
+      replaceState?: boolean
+    ) => {
+      const prevSearchString = urlSearchString;
+      const urlState = parseUrlState(prevSearchString);
+      const parsedQueryString = parse(prevSearchString, { sort: false });
+
+      if (!Object.prototype.hasOwnProperty.call(urlState, accessor)) {
+        urlState[accessor] = {};
+      }
+
+      if (typeof attribute === 'string') {
+        if (isEqual(getNestedProperty(urlState, `${accessor}.${attribute}`), value)) {
+          return prevSearchString;
+        }
+
+        urlState[accessor][attribute] = value;
+      } else {
+        const attributes = attribute;
+        Object.keys(attributes).forEach((a) => {
+          urlState[accessor][a] = attributes[a];
+        });
+      }
+
+      try {
+        const oldLocationSearchString = stringify(parsedQueryString, {
+          sort: false,
+          encode: false,
+        });
+
+        Object.keys(urlState).forEach((a) => {
+          if (isRisonSerializationRequired(a)) {
+            parsedQueryString[a] = encode(urlState[a]);
+          } else {
+            parsedQueryString[a] = urlState[a];
+          }
+        });
+        const newLocationSearchString = stringify(parsedQueryString, {
+          sort: false,
+          encode: false,
+        });
+
+        if (oldLocationSearchString !== newLocationSearchString) {
+          const newSearchString = stringify(parsedQueryString, { sort: false });
+          if (replaceState) {
+            history.replace({ search: newSearchString });
+          } else {
+            history.push({ search: newSearchString });
+          }
+        }
+      } catch (error) {
+        // eslint-disable-next-line no-console
+        console.error('Could not save url state', error);
+      }
+    },
+    [history, urlSearchString]
+  );
+
+  return (
+    <StateProvider value={{ searchString: urlSearchString, setUrlState }}>{children}</StateProvider>
+  );
+};
 
 export const useUrlState = (accessor: Accessor) => {
   const { searchString, setUrlState: setUrlStateContext } = useContext(aiopsUrlStateStore);
